// apps/frontend/src/api/generated/model/historyListDtoRoomsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * QuizDock API
 * API REST du builder de quiz et des restitutions
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryListDtoRoomsItemSessionsItem } from './historyListDtoRoomsItemSessionsItem';

export type HistoryListDtoRoomsItem = {
  /** @nullable */
  roomId: string | null;
  /** @nullable */
  name: string | null;
  hostName: string;
  startedAt: string;
  endedAt: string;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  playerCount: number;
  sessions: HistoryListDtoRoomsItemSessionsItem[];
};